+++ client/configure.js
@@ -1,0 +1,4 @@
+'use strict'
+
+require('./scss/configure.scss')
+require('./js/configure.js')