+++ client/index.js
@@ -7,10 +7,6 @@ switch (logic) {
     require('./scss/login.scss')
     require('./js/login.js')
     break
-  case 'configure':
-    require('./scss/configure.scss')
-    require('./js/configure.js')
-    break
   default:
     require('./node_modules/highlight.js/styles/tomorrow.css')
     require('./node_modules/simplemde/dist/simplemde.min.css')