+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Interactive setup
 - Auth: GitHub and Slack authentication providers are now available
 - Auth: LDAP authentication provider is now available
+- Logs: Support for the logging services: Bugsnag, Loggly, Papertrail and Rollbar
 
 ### Changed
 - Native Compilation Removal: Replaced farmhash with md5